Futurize for Py3  (#57)
* [py3] futurize(stage1): modernized py2 code w/o dependencies

Tests are passing.

* [py3] futurize(stage2): builtins, and utilities from `future`

This is to introduce more safe changes to support Python 3.
It is likely to result in tests failures.

* [py3] futurize(stage2): fix broken tests

* [py3] fix `TypeError: a bytes-like object is required, not 'str'`

Use `six.PY3` to force non-unicode strings in Python 2.

* [py3] Add `future` as a dependency

* [py3] Replace unnecessary `old_div` usage with `/`

* [py3] Replace `iteritems()` / `iterkeys()` -> `six.iteritems()` / `six.iterkeys()`

This is in response to CR feedback as using `items()`, and `keys()`
in Py2 will result in the creation of new arrays. Thus, this commit
will use proxy methods from `six` instead.

* [py3] fix linting errors due to import order

* [py3] remove unused `urllib` dependencies from `codecs.py`, and fix tests

* [py3] revert uses of `old_div` in favour of standard `/` divisor

* [py3] Revert change from `long` -> `int`

Use `long` for py2, and `int` for py3. Otherwise we will get an
integer overflow for py2 32bit.