Allow specifying hostname and ip via tags (#167)
* Allow specifying the hostname via a tag

This is useful for applications running inside containers where the
hostname is autogenerated and generally useless. In kubernetes, the
downward api can be used to pass in the node name actually running
the container.

Fixes #66

Signed-off-by: Jeff Schroeder <jeffschroeder@computer.org>

* Catch the `socket.error` exception when trying to get the hostname

This is the base exception for all errors in the socket library. Explicit
is better than implicit!

Signed-off-by: Jeff Schroeder <jeffschroeder@computer.org>

* Don't attempt to lookup the ip address if specified via a tag

This matches the functionality of jaeger-client-java after
jaegertracing/jaeger-client-java#371 was merged.

Refs: #166

Signed-off-by: Jeff Schroeder <jeffschroeder@computer.org>

* Remove ipv4_to_int since jaeger converts tags to strings

Signed-off-by: Jeff Schroeder <jeffschroeder@computer.org>