Fix handling of missing headers in the b3 codec (#215)
* Fix handling of missing headers in the b3 codec

The b3 codec uses .get() to extract headers and so does not throw a key
error if headers are missing.  However many of these headers are decoded
via "header_to_hex" which throws an exception on a None value.  Other
codecs handle the missing header case by returning None, update the b3
codec to follow this convention.

Signed-off-by: Caleb Howe <Caleb.S.Howe@gmail.com>