Add max_traceback_length as parameter
Tracebacks used to be truncated at the same point as any other
string tag. As tracebacks tend to be long, this often resulted in
traceback messages being cut off halfway. Now they have their own
length limit, which defaults to four times longer than the limit for
normal string tags and can be set when a tracer is created.

Signed-off-by: Qi Qi <piano.keys0@gmail.com>