Add stack traceback to error logs (#250)
* Add stack traceback to error logs

Previously, error logs only come with a tag of the form
"<traceback object at location>". Now they also report the actual
stack trace.

Resolves: #163

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Change traceback to be reported in stack

Previously, the traceback message was reported in error.object rather
than stack. The traceback message is now reported in the stack tag
of a error log, and it comes after the traceback object location.

In addition, two tests have been added to validate this behavior.

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Remove traceback location in report

The message 'traceback object at location' has been removed, so that
only the informative message about the stack trace is included in the
'stack' tag.

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Fix linting errors

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Use input traceback object instead of global state

Instead of reporting the stack trace of the globally stored traceback
object using traceback.format_exc, now report the stack trace of the
input traceback object `value`.

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Automatically catch exception using span in tests

Instead of manually logging an exception, use the context manager
properties of the span to automatically catch and log exceptions
when testing for traceback reporting behavior.

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Add max_traceback_length as parameter

Tracebacks used to be truncated at the same point as any other
string tag. As tracebacks tend to be long, this often resulted in
traceback messages being cut off halfway. Now they have their own
length limit, which defaults to four times longer than the limit for
normal string tags and can be set when a tracer is created.

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Remove traceback length from 'process' methods

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Remove pytest cache

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Nit, add back commas

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Fix traceback test

Signed-off-by: Qi Qi <piano.keys0@gmail.com>

* Add traceback length to config, spans, etc

Signed-off-by: Qi Qi <piano.keys0@gmail.com>