Add package prefix to imports in generated code (#313)
* Add package prefix to imports in generated code

Packages generated with thrift refers to each other as if they are top
level packages. This prevents imports from working when generated
packages are placed inside some other package. Currently jaeger_client
relies on hacking sys.path, but this doesn't work properly if the
library is packed inside an application built using PyInstaller.
Another way to get imports to work is to rewrite them after
generation. This can be done easily since the build process already
involves an awk script that performs a couple of replacements.

Signed-off-by: Andrey Nikitin <ethframe@gmail.com>

* Remove unused import

Signed-off-by: Andrey Nikitin <ethframe@gmail.com>